Improved error reporting, fixed bug in markdown tag, added test cases
In the process of improving the markdown tag's static error reporting I
discovered that in some cases the markdown tag would not use the correct
directory for finding files.  Fixing that required modifying the
implementation of the apply tag to also update the current template directory
in addition to the current context.