Remove backslash escaping in attributes
Backslash escaping is a destructive operation, so won't work because it makes
behavior dependent on the number of times a node is processed.  This commit
also preserves the ${} syntax if no splice existed.  This should fix the
problem once and for all.  If you need ${} to be preserved in the attribute
unescaped, then just don't bind anything to that tag.